Adding console messages and exceptions. (I don't wanna be blind!)

+++ src/views/Login.vue
@@ -37,9 +37,17 @@ export default {
         let apiURL = 'http://localhost:4000/user/login';
         axios.post(apiURL, this.user).then((res) => {
           
-          //store token into "userInformations"
-          localStorage.setItem('userInformations', JSON.stringify(res.data.token));
+          //Return messages 
+          console.log(res.data.message);
           
+          if(res.data.isPswrdMatches){ //If the user exists and the password matches.
+            //store the token (use_id with token) into "userInformations"
+            localStorage.setItem('userInformations', JSON.stringify(res.data.token));
+
+            //redirect into the homepage or whatever.
+            this.$router.push('/')
+          }
+
           this.user = {
             name: '',
             email: '',
